fix_dates.py
import yaml
import os

base_dir = os.path.dirname(os.path.abspath(__file__))
config_path = os.path.join(base_dir, "config.yaml")

with open(config_path, 'r', encoding='utf-8') as f:
    config = yaml.safe_load(f)

for cls in config.get('fixed_classes', []):
    cls['active_dates'] = "2026-03-01 至 2026-07-01"

with open(config_path, 'w', encoding='utf-8') as f:
    yaml.dump(config, f, allow_unicode=True, sort_keys=False)

print("Dates updated!")
