requirements.txt
PyYAML>=6.0
requests>=2.31.0
openai>=1.12.0
